lucid_school/requirements.txt
flask==3.0.3
werkzeug==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
reportlab==4.2.2
requests==2.32.3
